+++ crates/code_generation/src/lib.rs
@@ -1,15 +1,15 @@
 use cyclers::generate_cyclers;
+use execution::{generate_replayer_struct, generate_run_function};
 use perception_databases::generate_perception_databases;
 use proc_macro2::TokenStream;
 use quote::quote;
-use run::{generate_replayer_struct, generate_run_function};
 use source_analyzer::{cyclers::Cyclers, structs::Structs};
 use structs::generate_structs;
 
 mod accessor;
 pub mod cyclers;
+pub mod execution;
 pub mod perception_databases;
-pub mod run;
 pub mod structs;
 pub mod write_to_file;
